requirements.txt
alembic==1.6.2
aniso8601==9.0.1
astroid==2.5.6
click==7.1.2
Flask~=2.0.0
Flask-Migrate==2.7.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==3.36.0
greenlet==1.1.0
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow~=3.12.1
marshmallow-sqlalchemy==0.25.0
mccabe==0.6.1
pylint==2.8.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
six==1.16.0
SQLAlchemy~=1.4.15
toml==0.10.2
Werkzeug~=2.0.0
wrapt==1.12.1

PyJWT~=2.1.0